Melhorando o código e adicionando mais funções

+++ lojinhainha_app/models/vendas/vendas_delete.py
@@ -1,12 +1,15 @@
 from database.connection import get_db
 
 def excluir_venda(id):
-    # conecta no banco
     db = get_db()
     cursor = db.cursor()
 
-    # remove venda
-    cursor.execute("DELETE FROM vendas WHERE id = %s", (id,))
+    #  verifica se a venda existe antes de excluir
+    cursor.execute("SELECT id FROM vendas WHERE id = %s", (id,))
+    venda = cursor.fetchone()
 
-    db.commit()
+    if venda:
+        cursor.execute("DELETE FROM vendas WHERE id = %s", (id,))
+        db.commit()
+
     db.close()
